+++ src/main/java/roomescape/dao/preparedstatementcreator/FindReservationPreparedStatementCreator.java
@@ -1,18 +1,18 @@
-package roomescape.dao.preparedstatement;
+package roomescape.dao.preparedstatementcreator;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
 import org.springframework.jdbc.core.PreparedStatementCreator;
 
-public class DeleteReservationPreparedStatementCreator implements PreparedStatementCreator {
+public class FindReservationPreparedStatementCreator implements PreparedStatementCreator {
 
     private static final String SQL =
-            "DELETE FROM reservation WHERE id=?;";
+            "SELECT * FROM reservation WHERE id=?;";
 
     private final Long id;
 
-    public DeleteReservationPreparedStatementCreator(Long id) {
+    public FindReservationPreparedStatementCreator(Long id) {
         this.id = id;
     }
